backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography>=3.4.8,<42.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.35.0
httpx==0.25.2
